fix: 修复多模块与依赖库 API 不匹配的编译错误
## 问题
项目除 lumina-common/lumina-framework 外，agent-core/business-base/business-agent
均无法编译（agentscope/nacos/jsqlparser 等依赖 API 与代码假设不匹配），导致整个
项目无法构建与启动。

## 主要变更

### lumina-agent-core（适配 AgentScope 1.0.7 与 Nacos API）
- DefaultAgentExecutionEngine: Msg 包路径 model->message；agent.call 改传
  List<Msg>；temperature 用 doubleValue 适配 Double 参数
- ToolDefinitionToAgentToolAdapter: ToolResultBlock.success->text（success 不存在）
- ConfigLoader: 移除 @NacosConfigListener（属性 dataIds 不存在，且该注解不适用
  spring-cloud-alibaba-nacos-config）

### lumina-business-agent
- AuthController: JwtUtil.parseTokenToLoginUser 由静态调用改为注入实例
  （方法依赖 @Value 注入的 secretKey）
- AgentServiceImpl: Page.getPages() 返回 long，强转 int 适配 setPages(Integer)

### lumina-business-base
- WebMvcConfig: 类名 BaseWebMvcConfig->WebMvcConfig（与文件名一致）
- TenantLineInterceptor: 适配 JSqlParser 4.9（移除已删除的 SelectBody，Select
  直接 instanceof PlainSelect/SetOperationList）
- BaseToolProvider: roleIds 为 Long[]，new ArrayList<>() 改为 new Long[0]
- RoleServiceImpl: 补 import UserRoleDO
- AuthController: 同 agent 模块，JwtUtil 改注入实例
- pom.xml: 新增 agent-core 依赖（BaseToolProvider 使用 @AgentTool 注解）

### 构建
- 父 pom: 锁定 jsqlparser 4.9（避免 5.x 移除 SelectBody 破坏 TenantLineInterceptor）

## 验证
- mvn compile -DskipTests 全 8 模块 BUILD SUCCESS

## 统计数据
- 12 个文件变更
- 新增 47 行，删除 38 行

+++ lumina-modules/lumina-business-agent/src/main/java/io/lumina/agent/service/impl/AgentServiceImpl.java
@@ -164,7 +164,7 @@ public class AgentServiceImpl implements AgentService {
         pageResult.setTotal(doPage.getTotal());
         pageResult.setPageNum(pageNum);
         pageResult.setPageSize(pageSize);
-        pageResult.setPages(doPage.getPages());
+        pageResult.setPages((int) doPage.getPages());
 
         return pageResult;
     }
